requirements.txt
Flask==0.12.2

# Application server for both development and production.
gunicorn==19.7.0

# Testing and static analysis.
pytest==3.3.2
pytest-cov==2.5.1
flake8==3.5.0

# CLI.
Click==6.4

# Data and workers.
psycopg2==2.7.3.2
Flask-SQLAlchemy==2.3.1
redis==2.10.6
celery==4.1.0

# Forms.
Flask-WTF==0.14.2
WTForms-Components==0.9.9

# Utils.
fake-factory==0.5.7

# Extensions.
flask-debugtoolbar==0.10.1
Flask-Mail==0.9.1
Flask-Login==0.3.2
Flask-Limiter==0.9.5
